note view: notify updates on note property changes
We're notifying updates as soon as any of the text views change its text, which
is not optimal since we're notifying false updates every time the user changes
the selected note (since the text view text changes). Instead, this commit
changes to notify every time any of the properties of the note associated to
the view changes.